move BpAnimatedTabLayout to design_system components
